requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
chromadb>=0.4.18
pymupdf>=1.23.7
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pytest>=7.4.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
streamlit>=1.18.0 
